FYP finallllllllllllllllll/requirements.txt
streamlit
pandas
plotly
reportlab
kaleido
mysql-connector-python
pandas